requirements.txt
torch==2.3.0
torchvision==0.18.0
diffusers==0.33.1
transformers==4.51.3
accelerate==1.6.0
scikit-image==0.25.2
Pillow==10.3.0
matplotlib==3.8.4
tqdm==4.66.4
numpy==1.26.4
datasets==2.19.1
